internal: bptree: remove keys from internal nodes and leaves
After insertion operations, rightmost key from every leaf in the bptree
except the rightmost leaf will have a copy in an internal node pointing
to the subtree for which it is the rightmost key. However, if the
rightmost key from a leaf is deleted, the code doesn't switch the key in
the internal node to point to the new rightmost key. This means that
extra memory is used to store keys with no records in the tree. It also
causes a bug in _find() with LOOKUP_PREFIX, since _find_leaf() might
return a leaf block where all the matching keys have been deleted,
causing the lookup to fail when matching keys do exist in the next leaf.