internal: bptree: Fix leaf_node redistribution
When the entries are removed from the tree, _redistribute_nodes() was
setting bk_prime to the leftmost key of the right leaf, instead of the
rightmost key of the left leaf, which is what the rest of the code
assumes. This means the _find() code could no longer find the record
that matches bk_prime. Fix this.