ucmd: module: dm: add sid_ucmd_dm_subsys_match to match DM subsystem
Add sid_ucmd_dm_subsys_match which is DM-specific hook for DM submodules
to scan for, detect and claim a DM subsystem. The DM module iterates over
all its submodules and the first submodule that matches a device with
the subsystem that the submodule handles, claims the device. Once the
subsystem is resolved, the claim is store in the KV store so we don't
need to detect again on subsequent events.