resource: ucmd-module: add KV_NS_DEVMOD namespace
We already had KV_NS_MODULE for unique records in the scope of current
module and KV_NS_DEVICE for unique records in the scope of current
device.

The new KV_NS_DEVMOD combines KV_NS_DEVICE and KV_NS_MODULE so that we
can create records which are device-specific and at the same time
they are specific to a module. This way we don't need to explicitly
specify a predefined key prefix for a record which corresponds to the
module somehow and is bound to a device (for example, this can be
used for records like DM_NAME, DM_UUID, DM_MULTIPATH_VALID,
DM_MULTIPATH_WWID).

Taking the key schema: op:dom:ns:ns_part:id_cat:id:core, the
KV_NS_MODDEV is applied this way:

   ns      = "X"
   ns_part = <dev_id>
   id_cat  = "M"
   id      = <mod_name>