resource: kv-store: add kv store transactions
Add transactions to the kv store. During a transaction,
kv_store_set_value() and kv_store_unset() are handled differently.  when
setting a value, the _update_fn will save the old value in the rollback
buffer. When unsetting a value, the _unset_fn will not actually perform
the unset, but instead save the keys to be unset in the unset buffer.
When the transaction is ended, if it was successful, the saved keys will
be unset. If the transaction was unsuccessful, the old values will be
restored.