resource: worker-control: add worker_control_run_new_worker
The new worker_control_run_new_worker is used to execute external worker
while waiting for it to complete in situ. This way, we don't need to hook
into existing event loop which may reside several layers up the current
call stack. That means, we don't need to do current call stack unwinding
up to the event loop point and then returning back which would be quite
comples. Also, it would be useless if we need the external worker to
finish anyway for us to continue processing based on its results.

The worker_control_run_new_worker is a combination of
worker_control_get_new_worker and worker_control_run_worker (in the
worker) and running a new and temporary event loop (in the worker proxy)
to handle just this worker.