resource: worker-control: make it possible to set custom arg for worker and its proxy
Make it possible to set custom arg for worker and its proxy. Then,
whenerver we're using the worker and/or proxy, we are able to retrieve
the custom information directly. This is practical mainly if we're
processing worker/proxy related callbacks (e.g. channel rx/tx callbacks)
where we can pass the extra arg.

Note: the rx/tx worker/proxy channel callbacks already have their own
arg, but since the time when we define the channels (with their arg)
and the time when we actually create the worker/proxy can differ, the
arg for the channel callback and the worker/proxy as a whole can differ
too.