resource: kv-store: don't pack kv_rollback_arg
The rollback buffer is an array of kv_rollback_args. Packing them could
cause them to no longer be correctly aligned. Since size_t should always
be the same size as a pointer, this is mostly a theoretical problem, but
compilers may generate slower code to access members of packed
structures.