resource: worker-control: add worker_control_get_worker_state
The worker_control_get_worker_state is used to check current state of
the worker that the worker_proxy represents.

This is mostly useful if we need to know whether we have received whole
output from the worker or if it could have been interrupted due to a
timeout (the check can be added to worker_proxy's channel worker_rx_cb
callback).