resource: ubridge: fix KV index management while unsetting a record
When a KV record is unset (using _do_sid_ucmd_set_kv), we are actually setting
a void value inside the kv_scalar_t wrapper. For this "unset" operation
to get synced with the main KV store, we still need to mark the record with
dedicated index. The index management was skipped for KV unset operation.
Fix this.