resource: ubridge: make sure value members are aligned.
Moving the gennum structure to after the flags structure, and forcing
the flags structure to be a 64bit integer, ensures that all the members
of the structure will be aligned, as long as the structure itself is
pointer aligned. Since kv_store_value->data is pointer aligned, and
kv_vector_t is pointer aligned, all values from the kv store will be
correctly aligned.