resource: ubridge: simplify flagging persistent records
Former KV_PERSISTENT flag had meaning only when used together with KV_SYNC
(so for the record to be persistent, we also needed to sync it with the
main KV store first). Simplify the flag so we don't need to add "KV_SYNC"
all the time. Rename the former KV_PERSISTENT to KV_SYNC_P so it's clear
that it's actually a "sync" operation, just with the extra "also mark
as persistent" part.