resource: ucmd-module: support setting an archive for a key
Add 'KV_ARCHIVE' sid_ucmd_kv_flags_t for sid_ucmd_set_kv to also set an
archive for given key. The archive key is derived automatically by
adding KV_PREFIX_OP_ARCHIVE_C ('~') at the beginning of the original key.

If we're working with a snapshot copy of the main KV store (in worker
process), then the actual value of the archive from snapshot is not synced
with the main KV store. Instead, the main KV store is signalled by the
'KV_ARCHIVE' flag to do an archive of the actual value found at the time
of sync in the main KV store.