resource: ubridge: fix value member order when using VVALUE_HEADER_PREP_BUF
Commit 28ef8c77f69acca36a95cae0ec50add0ed00ff12 shuffled value header
members for the structure to be aligned and the VVALUE_HEADER_PREP
macro parameters were reordered accordingly, including the args for all
the the macro calls.

However, the parameters for VVALUE_HEADER_PREP_BUG were not reordered
while the arguments in its calls were. This caused the headers in the
records which were prepared by VVALUE_HEADER_PREP_BUF to have incorrect
seqnum and gennum assigned (the GIN/GMB records created by
_refresh_device_disk_hierarchy_from_sysfs). Fix this.