resource: kv-store: fix and clean up archive handling in kv_store_set_value_with_archive
Setting the value for the 'archive_key' uses exactly the same update
procedure as when setting the value for the 'key'. The only difference
is that we are not calling given 'kv_update_fn' callback function for
the archive case. Therefore, factor out the common part to a new
'_set_value' helper function and call this one instead.

Also, if setting the 'archive_key' fails and we need to rollback the
value for 'key' we set right before, reuse existing '_kv_rollback_value'
helper function we already use when handling a transaction. Call the
rollback only in case we're NOT under a transaction. If we are under
a transaction, the rollback will be automatically called with the
'kv_store_transaction_end'.