resource: kv-store: do not destroy archived value when handling rollbacks
If we have an archive key "AK" and a key "K" and if we are doing a
rollback, do not destroy the current value for "AK" because that
is/will be rolled back for "K".

(STEP 1)
For example, if at the beginning, we have:

  AK = "A"
  K  = "B"

(STEP 2)
Then after setting the K to "C", we have:

  AK = "A"  ---> AK = "B"
  K  = "B"  ---> K  = "C"

  rollback_buffer = [K = "B", AK = "A"]

===

(STEP 3 - ROLLBACK)
When rolling back:

  AK = "B"  --> AK = "A" (we CANNOT destroy "B" as it is the value rolled back for K)
  K  = "C"  --> K  = "B" (here is that "B" value we need to preserve)

===

(STEP 3 - NO ROLLBACK)
If NOT doing a rollback, we normally destroy all the values in the rollback buffer
as they are no longer needed. However, if the value is archived (like value "B"
from STEP 2), we CANNOT destroy that as it is still used for the archive, so we
do:

  skip_destroy(K = "B")
  destroy(AK = "A")