resource: ubridge: add _vvalue_header_prep and _vvalue_data_prep
The _vvalue_header_prep and _vvalue_data_prep are the same as the
existing VVALUE_HEADER_PREP and VVALUE_DATA_PREP macros. We will use
these functions to replace the macros since the logic inside got a bit
more complex now so it's not that feasible defining these as macros.