resource: ubridge: separate remove events from regular scans
Instead of using the regular scan phases for removes, use a separate
set, with just the same initialization and cleanup as a regular scan.
The main remove is done in two phases. The first calls the new module
function, scan_remove for the appropriate type and block modules.
These are currently all stubs. The second phase removes the core SID
kvstore values. Currently this just removes the device from the
device hierarchy in the kvstore.