resource: ubridge: run block modules first in ident phase
If SID fails to get the module name at all, it will not run the ident
functions of the block modules. However, it will run all the other scan
phase functions. In all these other scan phase functions, the block
modules will be run, regardless of whether the device has a type module,
even if the module name couldn't be determined.  Make the ident phase
also run the block modules, even if the module name can't be determined,
to be consistent.