resource: ubridge: fix freed memory access in _kv_delta_set
When _delta_step_calc() is building the delta buffers, it can add
entries to the delta->final and delta->minus buffers that point to data
stored in the old vvalue (delta->plus will only ever point to new data,
which isn't stored in the old vvalue). When _kv_delta_set finally writes
the new vvalue from delta->final, the old vvalue is freed. This means
that there can be entries in delta->final and delta->minus that point to
freed memory. The same thing can happen to abs_delta->plus and
abs_delta->minus after those are written out.

Where this is actually visible is when a device is removed and the
device hierarchy is updated. In this case, delta->minus ends up with
entries pointing to freed memory. If that memory is overwritten, it
will effect both the DELTA_WITH_DIFF and DELTA_WITH_REL output.

The easiest way to avoid this is to run _kv_delta_set() inside a
transaction, so the old values don't get freed until the end, when
the transaction is ended.