resource: support definition of extra custom log prefixes for resource types
The extra custom log prefix that is defined for a resource type is added
for all instances of that type right before their resource id prefix.

This can be used to either highlight certain log lines that belong to
handling certain resource types or to logically group logs from
different resource types (if they use the same custom log prefix).