resource: worker-control: add service_link_defs[] param to worker_control_run_worker and worker_control_run_new_worker
When a new worker is created, we need to be able to provide service link
definition array so that all required service links are established for the
worker too.

The reason is that the "worker resource" is created as a top level
resource in the worker process (so without having a parent) and hence
we can't inherit the service links from the parent anymore.