res: ubr: remove superfluous 'index' arg for internal kvs handling fns
We need the kvs indexing in snapshot only to quickly access records which
we need to send over to main kvs for syncing. Obviously, we don't need such
indexing for the main kvs itself.

To make a difference, we can use the new struct's kv_update_arg.is_sync
field to check this:

  - If handling the main kvs, hence "syncing", hence "no indexing".
    kv_update_arg.is_sync == true

  - If handling the snapshot kvs, hence "not syncing", hence "indexing".
    kv_update_arg.is_sync == false

The struct kv_update_arg is passed to all kvs handling functions,
so we can always check kv_update_arg.is_sync and directly deduce
a need for indexing.