res: ubr: pass both resource and owner string to internal fns handling KV records
A resource handler is now also used for logging, not only for making a
distinction between a module setting a KV record or the core itself.

Add "const char *owner" parameter to all KV record handling functions
that are used by both core and modules (usually named as "_do_sid_ucmd_*").
The "owner" values were normally acquired in the KV record handling
functions anyway, so this patch mostly moves this responsibility up
to the caller where we know for sure the function is called from the
module or core.