res: ubr: rework internal fn to set basic set of device KV records
Rework internal fn (_set_dev_kvs) that is called on each uevent to set
basic set of device KV records if they are not set already. Make it
clear when we are and from where we are importing the values (SID's KV
store or udev db). Also, prepare code for making sure that SID KV store
and udev db are in sync when it comes to basic device info: dev ID,
diskseq number, dev name and dev major:minor pair (this will be a part
of a future patch).