res: ubr: use '-partX' suffix to extend DISKSEQ for a partition
Partitions have the same DISKSEQ number assigned as their whole
device. Since we use this number to differentiate devices, including
partitions, we need to add suffix for the partitions. Take the 'PARTN'
variable from udev environment that is always set for partitions
and use that as the suffix with this scheme: <DISKSEQ-partPARTN>

This is the same scheme as udev already uses in its default udev
rules for items created under /dev/disk/by-diskseq.